Arduino Robot rev2
Rev2 is getting the robot dialed in onto the buttons, and holding the position and not continuing the code to return the robot to the home position. This is accomplished by only writing code in the setup section of the Arduino code. Nothing needs to be written in the loop. I also stopped moving the servos back to the home position as we just want to push the buttons for testing
With this version the arduino code needs to be reset everytime the program has finished. To bring the code back to the very beginning.